Intro to R: "gene_expression.csv"
The raw data was downloaded using the link provided in the github [1]. The data file was allocated to "a" and read the data making the gene accession number as the row names. The type of the data was converted to data frame[5]. A column was created to get the mean of values of the other two columns and sorted the data according to the given instructions ( mean values less than 10). Projecting the selected values in pairs plot graph.